backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
SQLAlchemy = "^1.4.29"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
PyMySQL = "^1.0.2"
alembic = "^1.7.5"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
